refactor: api for new features and fixes (#73)
refactor: api for new features and fixes

+++ api-service-leave-forcasting/src/main/java/com/ibsplc/apiserviceleaveforcasting/response/EmployeeLeaveReportResponse.java
@@ -15,5 +15,5 @@ public class EmployeeLeaveReportResponse {
     private String employeeId;
     private String employeeName;
     private List<MonthLeaveSummaryResponse> month;
-    private int noOfDays;
+    private double noOfDays;
 }
